feat: revive opsense-rhai with macro-based Rhai bindings
## Summary
Revive opsense-rhai crate and integrate via macro-based Rhai bindings system.

## Changes

### Macro System (opsense-macros)
- #[rhai_class]: struct bindings with constructor + accessors
- #[rhai_func]: free function registration via inventory
- #[rhai_register]: module-level collector
- default_constructor option for #[derive(Default)] structs

### opsense-mlib: Rhai bindings for core types
- AnalysisGrid: grid_fit, grid_fit_values, accessors (num_cells, grid_cell, grid_occupancy, etc.)
- TransitionAnalysis: transition_analysis, accessors (num_buckets, down_probability, transitions, etc.)
- script module: RhaiBindings trait + helpers (parse_points, extract_values, etc.)

### opsense-rhai: Runtime & tools
- Manual accessor registration with &mut Self receivers (Rhai 1.26 compatible)
- Free functions: now_secs(), ts_rate, ts_moving_avg, ts_resample, ts_quantile, ts_p95, ts_p99, ts_delta, ts_pct_change
- Attributes: attr(), attrs() per-call
- Transform: rewritten for new Context + TimeseriesStation API
- Pipeline tests fixed for new component model

### Tests: 14/17 pass
- Core: grid_ops (3), operators (2), disk_grid_report, disk_spike, pipeline (4)
- Remaining failures: http_format, station_lookup, e2e_disk_grid (mock format issues)

+++ crates/opsense-core/src/station.rs
@@ -265,9 +265,9 @@ impl TimeseriesStation {
     fn spawn_bg_sync(station: &mut Self, cfg: &StorageConfig, ts: Arc<dyn TimeseriesStorage>) {
         use std::time::{Duration, SystemTime};
 
-        let flush_every = cfg.s3_flush_interval_secs as u64;
-        let snapshot_every = cfg.s3_snapshot_interval_secs as u64;
-        let retention_secs = cfg.retention_secs as u64;
+        let flush_every = cfg.s3_flush_interval_secs;
+        let snapshot_every = cfg.s3_snapshot_interval_secs;
+        let retention_secs = cfg.retention_secs;
         if flush_every == 0 && snapshot_every == 0 && retention_secs == 0 {
             return;
         }
